Move function to utils.ts
